pyproject.toml
[project]
name = "AA84-API"
version = "0.1.0"
description = "AA84 API"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic[email]",
    "python-multipart",
    "python-dotenv",
    "pydantic-settings",
    "sqlalchemy",
    "psycopg2-binary",
    "bcrypt",
    "pyjwt",
    "generate-password",
]